+++ lib/ood_core/job/status.rb
@@ -72,7 +72,7 @@ module OodCore
       end
 
       # Generate a hash value for this object
-      # @return [Fixnum] hash value of object
+      # @return [Integer] hash value of object
       def hash
         [self.class, to_sym].hash
       end