+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Basic multi-cluster support for LSF by specifying name of cluster for -m
+  argument. [#24](https://github.com/OSC/ood_core/issues/24)
 
 ## [0.2.1] - 2018-01-26
 ### Changed