+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.3.0] - 2018-04-05
 ### Added
 - Basic multi-cluster support for LSF by specifying name of cluster for -m
   argument. [#24](https://github.com/OSC/ood_core/issues/24)
@@ -110,7 +112,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Initial release!
 
-[Unreleased]: https://github.com/OSC/ood_core/compare/v0.2.1...HEAD
+[Unreleased]: https://github.com/OSC/ood_core/compare/v0.3.0...HEAD
+[0.3.0]: https://github.com/OSC/ood_core/compare/v0.2.1...v0.3.0
 [0.2.1]: https://github.com/OSC/ood_core/compare/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/OSC/ood_core/compare/v0.1.1...v0.2.0
 [0.1.1]: https://github.com/OSC/ood_core/compare/v0.1.0...v0.1.1